requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
bertopic
corextopic
scikit-learn
sentence-transformers
Pillow
wordcloud
nltk
faiss-cpu
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv




